metabase: drop container volumes bucket
This is the last bucket with object information that was not included to the
Metadata bucket. This commit also changes how metabase treats object counters
for outside reports: _every_ object type is now calculated, not only REGULAR
ones.

+++ pkg/local_object_storage/shard/delete.go
@@ -64,7 +64,7 @@ func (s *Shard) deleteObjs(addrs []oid.Address) error {
 	for i := range res.RemovedObjects {
 		removedPayload := res.RemovedObjects[i].PayloadLen
 		totalRemovedPayload += removedPayload
-		s.addToContainerSize(res.RemovedObjects[i].Address.Container().EncodeToString(), -int64(removedPayload))
+		s.addToContainerSize(res.RemovedObjects[i].Address.Container().EncodeToString(), -int(removedPayload))
 
 		err = s.blobStor.Delete(res.RemovedObjects[i].Address)
 		if err == nil {
@@ -80,7 +80,7 @@ func (s *Shard) deleteObjs(addrs []oid.Address) error {
 		}
 	}
 
-	s.addToPayloadCounter(-int64(totalRemovedPayload))
+	s.addToPayloadCounter(-int(totalRemovedPayload))
 
 	return nil
 }
